Issue #247 - Very minor refactor to suppress a warning - signal correction (read comment)
If you name a slot matching a pattern "on_*" in a QObject class, Qt will
try to connect that to a matching signal.

It's useful for GUI code because it is generated by Qt, not so much for
non GUI code.

This causes warnings in the console output like
"QMetaObject::connectSlotsByName: No matching signal for
on_shutterCloseTimer_timeout()", however, your code will work as
expected.

Message: don't name slots "on_..." if it's not GUI code.

See
https://stackoverflow.com/questions/43918098/signal-to-slot-in-qt-creator-where-is-connect-function